daredevil section is wip
