+++ components/1_Header/header.scss
@@ -15,6 +15,7 @@ header {
 		background-position: center top;
 		background-repeat: no-repeat;
 		background-size: cover;
+		background-attachment: fixed;
 		z-index: 1;
 	}
