+++ components/3_Banner_Daredevil/Banner_daredevil.js
@@ -1,0 +1,13 @@
+import React, {Component} from 'react';
+import style from './Banner_daredevil.scss'
+
+class BannerDaredevil extends Component {
+  render() {
+    return (<section className={style.bannerDaredevil}>
+      <p>Daredevil</p>
+      <figure></figure>
+    </section>)
+  }
+}
+
+export default BannerDaredevil;
