+++ components/3_Banner_Daredevil/Banner_daredevil.scss
@@ -1,0 +1,31 @@
+.bannerDaredevil {
+  width: 100vw;
+  height: 350px;
+  background: #000000;
+  display: grid;
+  grid-template-columns:
+    [c1] 1fr [c2] 350px [c3];
+  grid-template-rows:
+    [r1] 1fr [r2];
+
+  p {
+    grid-column: [c1]/[c2];
+    grid-row: [r1]/r2;
+    justify-self: center;
+    align-self: center;
+    font-size: 15rem;
+    color: #E32800;
+    text-transform: uppercase;
+  }
+
+  figure {
+    width: 100%;
+    height: 100%;
+    grid-column: [c2]/[c3];
+    grid-row: [r1]/r2;
+    background-image: url("./img/mathew.jpg");
+    background-repeat: no-repeat;
+    background-position: center center;
+    background-size: cover;
+  }
+}
