+++ components/4_Daredevil/daredevil.scss
@@ -1,21 +1,24 @@
-section {
+.daredevilContent {
 	width: 100vw;
 	height: 100vh;
-	background: url('./img/daredevil.jpg');
+	background: url("./img/daredevil.jpg");
 	background-position: center center;
 	background-size: cover;
+	background-attachment: fixed;
 	display: grid;
 	grid-template-columns:
 		[c1] 150px [c2] 640px [c3] 1fr [c4];
 	grid-template-rows:
 		[r1] 200px [r2] 360px [r3] 1fr [r4];
+
 	span {
 		width: 100%;
 		height: 100%;
 		grid-column: [c2]/[c3];
 		grid-row: [r2]/[c3];
+
 		iframe {
-			box-shadow: 5px 5px 80px #E32800;
+			box-shadow: 0 0 100px #E32800;
 		}
 	}
 }
